# Improve `Labels.Compare()` performance with `stringlabels`