@colega 2024-03-18
After figuring out that this most likely is because I've removed one jump condition, I decided to try to remove it from the previous loop too (when checking last bytes). I've added an `and()` function that prevents the short-circuit return (and gets inlined anyway), and ran the benchmarks again for all versions with count=25. This new one is `new_and`:
```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/model/labels
                                       │   main_25   │               new_25                │             new_and_25              │
                                       │   sec/op    │   sec/op     vs base                │   sec/op     vs base                │
Labels_Compare/equal                     5.764n ± 1%   5.814n ± 2%        ~ (p=0.560 n=25)   5.046n ± 0%  -12.46% (p=0.000 n=25)
Labels_Compare/not_equal                 11.73n ± 1%   11.06n ± 0%   -5.71% (p=0.000 n=25)   11.21n ± 0%   -4.43% (p=0.000 n=25)
Labels_Compare/different_sizes           4.999n ± 1%   4.947n ± 1%   -1.04% (p=0.046 n=25)   4.548n ± 0%   -9.02% (p=0.000 n=25)
Labels_Compare/lots                      20.99n ± 1%   17.60n ± 1%  -16.15% (p=0.000 n=25)   18.04n ± 0%  -14.05% (p=0.000 n=25)
Labels_Compare/real_long_equal           15.00n ± 0%   14.97n ± 0%        ~ (p=0.377 n=25)   14.73n ± 1%   -1.80% (p=0.000 n=25)
Labels_Compare/real_long_different_end   25.25n ± 1%   24.38n ± 0%   -3.45% (p=0.000 n=25)   24.59n ± 0%   -2.61% (p=0.000 n=25)
geomean                                  11.79n        11.26n        -4.46%                  10.90n        -7.52%
```

Of course, the result will depend more on the length of the specific `data` than on the testcase itself, but IMO this change is worth the improvement.